refactor: fcm 로직을 메인 페이지로 이동
