+++ src/components/common/SideBar/SideBar.jsx
@@ -17,75 +17,25 @@ import firebase from 'firebase/compat/app';
 import Alarm from '@/icons/alarm.svg';
 import AlarmOff from '@/icons/alarm_off.svg';
 import 'firebase/compat/messaging';
-import { useAlarm, useStoreId } from '@/store/useStoreId';
-import { useMutation } from '@tanstack/react-query';
-import postFcm from './apis/postFcm';
-import postUserFcm from './apis/postUserFcm';
+import { useAlarm } from '@/store/useStoreId';
+import { firebaseConfig } from '@/utils/firebase/firebase';
 
 const cn = classNames.bind(styles);
 
-const firebaseConfig = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
-  measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID,
-};
-
 export default function SideBar() {
-  const alarmSoundRef = useRef(null);
   const [canPlaySound, setCanPlaySound] = useState(true);
   const pathName = usePathname();
   const { setAlarm } = useAlarm();
-  const token = localStorage.getItem('token');
-  const { storeId } = useStoreId();
-
-  const userFcmMutation = useMutation({
-    // @ts-ignore
-    mutationFn: () => postUserFcm(token, storeId),
-  });
-
-  const fcmMutation = useMutation({
-    // @ts-ignore
-    mutationFn: (param) => postFcm(param),
-    onSuccess: () => {
-      // @ts-ignore
-      userFcmMutation.mutate();
-    },
-  });
+  const alarmSoundRef = useRef(null);
 
   if (!firebase.apps.length) {
-    console.log('Initializing Firebase...');
     firebase.initializeApp(firebaseConfig);
-  } else {
-    console.log('Firebase already initialized');
   }
 
   const messaging = firebase.messaging();
 
   useEffect(() => {
-    alarmSoundRef.current = new Audio('/mp3/alarm.mp3');
-  }, []);
-
-  useEffect(() => {
-    Notification.requestPermission().then((permission) => {
-      if (permission === 'granted') {
-        console.log('Notification permission granted.');
-        messaging
-          .getToken()
-          .then((fcmToken) => {
-            // @ts-ignore
-            fcmMutation.mutate({ token: token, fcmToken: fcmToken });
-          })
-          .catch((error) => {
-            console.error('Error getting token:', error);
-          });
-      } else {
-        console.log('Unable to get permission to notify.');
-      }
-    });
+    alarmSoundRef.current = new Audio('/mp3/mayo-alarm.mp3');
 
     if ('serviceWorker' in navigator) {
       navigator.serviceWorker
@@ -98,7 +48,6 @@ export default function SideBar() {
         });
     }
 
-    // 메시지 수신 처리
     messaging.onMessage((payload) => {
       if (canPlaySound) {
         alarmSoundRef.current.play().catch((error) => {
